Configure model output

Configure model output

vs2dh.ConfigureBasicOutput(times = vs2dh.ConfigureBasicOutputTimes(),
  main = vs2dh.ConfigureBasicOutputMain(), variables = TRUE,
  balance = vs2dh.ConfigureBalance(),
  obsPoints = vs2dh.ConfigureObsPoints(),
  boundaryFluxes = vs2dh.ConfigureBoundaryFluxes(),
  enhancedPrecision = FALSE)

Arguments

times
as retrieved by vs2dh.ConfigureBasicOutputTimes()

main
as retrieved by vs2dh.ConfigureBasicOutputMain()

variables
If TRUE output of pressure heads (and temperatures if TRANS = T) to file "variables.out" is desired at selected observation times (default: TRUE)

balance
IF TRUE one-line mass balance summary for each time step is written to file "balance.out"; (default: TRUE)

obsPoints
as retrieved by vs2dh.ConfigureObsPoints(), default: no output of observation points to file "obsPoints.out"

boundaryFluxes
as retrieved by vs2dh.ConfigureBoundaryFluxes(), default: no output of boundary fluxes to file "boundaryFluxes.out"

enhancedPrecision
If TRUE results are written with enhanced precision to files9 ("variables.out") and file11 ("obsPoints.out"), (default: FALSE)

Value

Configuration of model output