Basic solver configuration

Basic solver configuration

vs2dh.ConfigureBasicSolver(cis = TRUE, cit = TRUE, numt = 1000,
  minit = 2, itmax = 80, eps = 1e-04, eps1 = 0.001, eps2 = 0.001,
  hmax = 0.7, itstop = TRUE)

Arguments

cis
If TRUE spatial discretisation is realised by centered-in-space differencing; if FALSE backward-in-space differencing is to be used for transport equation. (default: TRUE)

cit
If TRUE temporal discretisation is realised by centered-in-time differencing; if FALSE backward-in-time or fully implicit differencing is to be used. (default: TRUE)

numt
Maximum number of time steps.(default: 1000). (NOTE: if enhanced precision in print out to file "balance.out" and file 11 "obsPoints.out", is desired set NUMT equal to a negative number. That is, multiply actual maximum number of time steps by -1)1

minit
Minimum number of iterations per time step. (default: 2)

itmax
Maximum number of iterations per time step. (default: 80)

eps
Head closure criterion for iterative solution of flow equation, L. (default: 0.0001)

eps1
Temperature closure criterion for iterative solution of transport equation, C. (default: 0.001)

eps2
Velocity closure criterion for outer iteration loop at each time step, L/T. (default: 0.001)

hmax
Relaxation parameter for iterative solution. See discussion in Lappala and others (1987) for more detail. Value is generally in the range of 0.4 to 1.2. (default: 0.7)

itstop
If TRUE simulation is terminated after ITMAX iterations in one time step; otherwise = F. (default: TRUE)

Value

Configuration of basic solver