Configure boundary conditions

Configure boundary conditions

vs2dh.ConfigureBoundaryCondition(jj = 5, nn = 17:26, ntx = 2,
  pfdum = 0.009778035, ntc = 0, cf = 10, importData = NULL,
  ibc = 0)

Arguments

jj
vector with row numbers of nodes

nn
vector with column numbers of nodes

ntx
vector with node type identifier for boundary conditions. 0 (for no specified boundary (needed for resetting some nodes after initial recharge period); 1 (for specified pressure head); 2 (for specified flux per unit horizontal surface area in units of L/T); 3 (for possible seepage face); 4 (for specified total head); 5 (for evaporation, Note: is not implemented yet!); 6 (for specified volumetric flow in units of L3/T). 7 (for gravity drain). (The gravity drain boundary condition allows gravity driven vertical fow out of the domain assuming a unit vertical hydraulic gradient. Flow into the domain cannot occur.)"

pfdum
vector with specified head for NTX = 1 or 4 or specified flux for NTX = 2 or 6. If codes 0, 3, 5, or 7 are specified, the line should contain a dummy value for PFDUM or should be terminated after NTX by a blank and a slash (/).

ntc
vector with node type identifier for transport boundary conditions. 0 (for no specified boundary); 1 (for specified temperatures)"

cf
vector with specified temperature for NTC = 1 or NTX = 1, 2, 4, 6, or 7. Present only if TRANS = T.

importData
Optionally a data.frame with the colnames (jj,nn,ntx,pfdum,ntc,cf) can be used. In this case the input for the function parameters (jj,nn,ntx,pfdum,ntc,cf) will be ignored!

ibc
Code for reading in boundary conditions: 0 (by individual node), 1 (by row/column). (default: 0), Note: currently only option 0 is implemented

Value

Boundary conditions