Read VS2dh model output file with fluxes at boundary faces

Read VS2dh model output file with fluxes at boundary faces

vs2dh.readBoundaryFluxes(model.path, fileName = "boundaryFluxes.out",
  dbg = TRUE)

Arguments

model.path
full path to folder containing model output files

fileName
name of file containing fluxes at boundary faces (default: "boundaryFluxes.out")

dbg
if true text output

Value

read model results of boundary faces are imported in a R object

Examples

### Location of example vs2dh model contained in "kwb.vs2dh package"
model.path <- system.file("extdata", "vs2dh_example/tutorial2", package = "kwb.vs2dh")
vs2dh.readBoundaryFluxes(model.path)

#>             TIME bID WATER_FLUX_TOTAL WATER_FLUX_TIMESTEP WATER_FLUX_RATE
#> 1   1.200000E-04   4        2.713E-18           2.713E-18       2.261E-14
#> 2   1.200000E-04   1        0.000E+00           0.000E+00       0.000E+00
#> 3   2.640000E-04   4        8.261E-19          -1.887E-18      -1.310E-14
#> 4   2.640000E-04   1        0.000E+00           0.000E+00       0.000E+00
#> 5   4.368000E-04   4       -2.447E-18          -3.273E-18      -1.894E-14
#> 6   4.368000E-04   1        0.000E+00           0.000E+00       0.000E+00
#> 7   6.441600E-04   4       -1.545E-17          -1.300E-17      -6.272E-14
#> 8   6.441600E-04   1        0.000E+00           0.000E+00       0.000E+00
#> 9   8.929920E-04   4       -2.136E-17          -5.913E-18      -2.376E-14
#> 10  8.929920E-04   1        0.000E+00           0.000E+00       0.000E+00
#> 11  1.191590E-03   4       -1.994E-17           1.428E-18       4.782E-15
#> 12  1.191590E-03   1        0.000E+00           0.000E+00       0.000E+00
#> 13  1.549908E-03   4       -2.330E-17          -3.360E-18      -9.377E-15
#> 14  1.549908E-03   1        0.000E+00           0.000E+00       0.000E+00
#> 15  1.979890E-03   4       -3.755E-17          -1.426E-17      -3.316E-14
#> 16  1.979890E-03   1        0.000E+00           0.000E+00       0.000E+00
#> 17  2.495868E-03   4       -4.539E-17          -7.837E-18      -1.519E-14
#> 18  2.495868E-03   1        0.000E+00           0.000E+00       0.000E+00
#> 19  3.115042E-03   4       -4.077E-17           4.619E-18       7.460E-15
#> 20  3.115042E-03   1        0.000E+00           0.000E+00       0.000E+00
#> 21  3.858050E-03   4       -4.775E-17          -6.979E-18      -9.393E-15
#> 22  3.858050E-03   1        0.000E+00           0.000E+00       0.000E+00
#> 23  4.749660E-03   4       -6.567E-17          -1.792E-17      -2.010E-14
#> 24  4.749660E-03   1        0.000E+00           0.000E+00       0.000E+00
#> 25  5.819592E-03   4       -1.037E-16          -3.802E-17      -3.554E-14
#> 26  5.819592E-03   1        0.000E+00           0.000E+00       0.000E+00
#> 27  7.103511E-03   4       -2.456E-16          -1.419E-16      -1.105E-13
#> 28  7.103511E-03   1        0.000E+00           0.000E+00       0.000E+00
#> 29  7.203511E-03   4       -2.597E-16          -1.411E-17      -1.411E-13
#> 30  7.203511E-03   1        0.000E+00           0.000E+00       0.000E+00
#> 31  7.323511E-03   4       -2.867E-16          -2.703E-17      -2.252E-13
#> 32  7.323511E-03   1        0.000E+00           0.000E+00       0.000E+00
#> 33  7.467511E-03   4       -3.078E-16          -2.102E-17      -1.460E-13
#> 34  7.467511E-03   1        0.000E+00           0.000E+00       0.000E+00
#> 35  7.640311E-03   4       -3.479E-16          -4.017E-17      -2.325E-13
#> 36  7.640311E-03   1        0.000E+00           0.000E+00       0.000E+00
#> 37  7.847671E-03   4       -4.003E-16          -5.239E-17      -2.527E-13
#> 38  7.847671E-03   1        0.000E+00           0.000E+00       0.000E+00
#> 39  8.096503E-03   4       -4.735E-16          -7.319E-17      -2.941E-13
#> 40  8.096503E-03   1        0.000E+00           0.000E+00       0.000E+00
#> 41  8.395101E-03   4       -5.684E-16          -9.491E-17      -3.178E-13
#> 42  8.395101E-03   1        0.000E+00           0.000E+00       0.000E+00
#> 43  8.753419E-03   4       -7.508E-16          -1.823E-16      -5.089E-13
#> 44  8.753419E-03   1        0.000E+00           0.000E+00       0.000E+00
#> 45  9.183401E-03   4       -1.079E-15          -3.284E-16      -7.637E-13
#> 46  9.183401E-03   1        0.000E+00           0.000E+00       0.000E+00
#> 47  9.699379E-03   4       -1.671E-15          -5.918E-16      -1.147E-12
#> 48  9.699379E-03   1        0.000E+00           0.000E+00       0.000E+00
#> 49  1.031855E-02   4       -2.904E-15          -1.233E-15      -1.991E-12
#> 50  1.031855E-02   1        0.000E+00           0.000E+00       0.000E+00
#> 51  1.106156E-02   4       -5.914E-15          -3.010E-15      -4.051E-12
#> 52  1.106156E-02   1        0.000E+00           0.000E+00       0.000E+00
#> 53  1.116156E-02   4       -6.350E-15          -4.355E-16      -4.355E-12
#> 54  1.116156E-02   1        0.000E+00           0.000E+00       0.000E+00
#> 55  1.128156E-02   4       -6.918E-15          -5.687E-16      -4.739E-12
#> 56  1.128156E-02   1        0.000E+00           0.000E+00       0.000E+00
#> 57  1.142556E-02   4       -7.674E-15          -7.555E-16      -5.247E-12
#> 58  1.142556E-02   1        0.000E+00           0.000E+00       0.000E+00
#> 59  1.159836E-02   4       -8.697E-15          -1.023E-15      -5.923E-12
#> 60  1.159836E-02   1        0.000E+00           0.000E+00       0.000E+00
#> 61  1.180572E-02   4       -1.012E-14          -1.423E-15      -6.865E-12
#> 62  1.180572E-02   1        0.000E+00           0.000E+00       0.000E+00
#> 63  1.205455E-02   4       -1.217E-14          -2.050E-15      -8.239E-12
#> 64  1.205455E-02   1        0.000E+00           0.000E+00       0.000E+00
#> 65  1.235315E-02   4       -1.522E-14          -3.046E-15      -1.020E-11
#> 66  1.235315E-02   1        0.000E+00           0.000E+00       0.000E+00
#> 67  1.271147E-02   4       -1.999E-14          -4.775E-15      -1.333E-11
#> 68  1.271147E-02   1        0.000E+00           0.000E+00       0.000E+00
#> 69  1.314145E-02   4       -2.780E-14          -7.809E-15      -1.816E-11
#> 70  1.314145E-02   1        0.000E+00           0.000E+00       0.000E+00
#> 71  1.365743E-02   4       -4.142E-14          -1.361E-14      -2.639E-11
#> 72  1.365743E-02   1        0.000E+00           0.000E+00       0.000E+00
#> 73  1.427660E-02   4       -6.690E-14          -2.549E-14      -4.116E-11
#> 74  1.427660E-02   1        0.000E+00           0.000E+00       0.000E+00
#> 75  1.501961E-02   4       -1.183E-13          -5.141E-14      -6.919E-11
#> 76  1.501961E-02   1        0.000E+00           0.000E+00       0.000E+00
#> 77  1.591122E-02   4       -2.303E-13          -1.120E-13      -1.256E-10
#> 78  1.591122E-02   1        0.000E+00           0.000E+00       0.000E+00
#> 79  1.698115E-02   4       -4.931E-13          -2.628E-13      -2.456E-10
#> 80  1.698115E-02   1        0.000E+00           0.000E+00       0.000E+00
#> 81  1.826507E-02   4       -1.152E-12          -6.589E-13      -5.132E-10
#> 82  1.826507E-02   1        0.000E+00           0.000E+00       0.000E+00
#> 83  1.980577E-02   4       -2.899E-12          -1.747E-12      -1.134E-09
#> 84  1.980577E-02   1        0.000E+00           0.000E+00       0.000E+00
#> 85  2.165462E-02   4       -7.738E-12          -4.839E-12      -2.617E-09
#> 86  2.165462E-02   1        0.000E+00           0.000E+00       0.000E+00
#> 87  2.387323E-02   4       -2.159E-11          -1.385E-11      -6.243E-09
#> 88  2.387323E-02   1        0.000E+00           0.000E+00       0.000E+00
#> 89  2.653556E-02   4       -6.220E-11          -4.061E-11      -1.525E-08
#> 90  2.653556E-02   1        0.000E+00           0.000E+00       0.000E+00
#> 91  2.973036E-02   4       -1.833E-10          -1.211E-10      -3.791E-08
#> 92  2.973036E-02   1        0.000E+00           0.000E+00       0.000E+00
#> 93  3.356412E-02   4       -5.496E-10          -3.663E-10      -9.553E-08
#> 94  3.356412E-02   1        0.000E+00           0.000E+00       0.000E+00
#> 95  3.816463E-02   4       -1.671E-09          -1.121E-09      -2.437E-07
#> 96  3.816463E-02   1        0.000E+00           0.000E+00       0.000E+00
#> 97  4.368525E-02   4       -5.147E-09          -3.476E-09      -6.297E-07
#> 98  4.368525E-02   1        0.000E+00           0.000E+00       0.000E+00
#> 99  5.030998E-02   4       -1.607E-08          -1.092E-08      -1.649E-06
#> 100 5.030998E-02   1        0.000E+00           0.000E+00       0.000E+00
#> 101 5.825967E-02   4       -1.151E-07          -9.903E-08      -1.246E-05
#> 102 5.825967E-02   1        0.000E+00           0.000E+00       0.000E+00
#> 103 6.779929E-02   4       -2.932E-05          -2.920E-05      -3.061E-03
#> 104 6.779929E-02   1        0.000E+00           0.000E+00       0.000E+00
#> 105 7.924684E-02   4       -1.077E-03          -1.048E-03      -9.152E-02
#> 106 7.924684E-02   1        0.000E+00           0.000E+00       0.000E+00
#> 107 9.298389E-02   4       -9.682E-03          -8.605E-03      -6.264E-01
#> 108 9.298389E-02   1        0.000E+00           0.000E+00       0.000E+00
#> 109 1.000000E-01   4       -1.771E-02          -8.026E-03      -1.144E+00
#> 110 1.000000E-01   1        0.000E+00           0.000E+00       0.000E+00
#> 111 1.001200E-01   4       -1.785E-02          -1.396E-04      -1.163E+00
#> 112 1.001200E-01   1        0.000E+00           0.000E+00       0.000E+00
#> 113 1.002640E-01   4       -1.802E-02          -1.702E-04      -1.182E+00
#> 114 1.002640E-01   1        0.000E+00           0.000E+00       0.000E+00
#> 115 1.004368E-01   4       -1.823E-02          -2.078E-04      -1.202E+00
#> 116 1.004368E-01   1        0.000E+00           0.000E+00       0.000E+00
#> 117 1.006442E-01   4       -1.848E-02          -2.540E-04      -1.225E+00
#> 118 1.006442E-01   1        0.000E+00           0.000E+00       0.000E+00
#> 119 1.008930E-01   4       -1.879E-02          -3.113E-04      -1.251E+00
#> 120 1.008930E-01   1        0.000E+00           0.000E+00       0.000E+00
#> 121 1.011916E-01   4       -1.917E-02          -3.824E-04      -1.281E+00
#> 122 1.011916E-01   1        0.000E+00           0.000E+00       0.000E+00
#> 123 1.015499E-01   4       -1.964E-02          -4.712E-04      -1.315E+00
#> 124 1.015499E-01   1        0.000E+00           0.000E+00       0.000E+00
#> 125 1.019799E-01   4       -2.023E-02          -5.828E-04      -1.355E+00
#> 126 1.019799E-01   1        0.000E+00           0.000E+00       0.000E+00
#> 127 1.024959E-01   4       -2.095E-02          -7.239E-04      -1.403E+00
#> 128 1.024959E-01   1        0.000E+00           0.000E+00       0.000E+00
#> 129 1.031150E-01   4       -2.185E-02          -9.033E-04      -1.459E+00
#> 130 1.031150E-01   1        0.000E+00           0.000E+00       0.000E+00
#> 131 1.038581E-01   4       -2.299E-02          -1.133E-03      -1.525E+00
#> 132 1.038581E-01   1        0.000E+00           0.000E+00       0.000E+00
#> 133 1.047497E-01   4       -2.442E-02          -1.429E-03      -1.603E+00
#> 134 1.047497E-01   1        0.000E+00           0.000E+00       0.000E+00
#> 135 1.058196E-01   4       -2.623E-02          -1.812E-03      -1.694E+00
#> 136 1.058196E-01   1        0.000E+00           0.000E+00       0.000E+00
#> 137 1.071035E-01   4       -2.854E-02          -2.313E-03      -1.801E+00
#> 138 1.071035E-01   1        0.000E+00           0.000E+00       0.000E+00
#> 139 1.086442E-01   4       -3.151E-02          -2.969E-03      -1.927E+00
#> 140 1.086442E-01   1        0.000E+00           0.000E+00       0.000E+00
#> 141 1.104931E-01   4       -3.534E-02          -3.831E-03      -2.072E+00
#> 142 1.104931E-01   1        0.000E+00           0.000E+00       0.000E+00
#> 143 1.127117E-01   4       -4.031E-02          -4.964E-03      -2.237E+00
#> 144 1.127117E-01   1        0.000E+00           0.000E+00       0.000E+00
#> 145 1.153740E-01   4       -4.676E-02          -6.451E-03      -2.423E+00
#> 146 1.153740E-01   1        0.000E+00           0.000E+00       0.000E+00
#> 147 1.185688E-01   4       -5.515E-02          -8.391E-03      -2.626E+00
#> 148 1.185688E-01   1        0.000E+00           0.000E+00       0.000E+00
#> 149 1.224026E-01   4       -6.604E-02          -1.089E-02      -2.840E+00
#> 150 1.224026E-01   1        0.000E+00           0.000E+00       0.000E+00
#> 151 1.270031E-01   4       -8.009E-02          -1.406E-02      -3.055E+00
#> 152 1.270031E-01   1        0.000E+00           0.000E+00       0.000E+00
#> 153 1.325237E-01   4       -9.807E-02          -1.798E-02      -3.257E+00
#> 154 1.325237E-01   1        0.000E+00           0.000E+00       0.000E+00
#> 155 1.391484E-01   4       -1.208E-01          -2.270E-02      -3.426E+00
#> 156 1.391484E-01   1        0.000E+00           0.000E+00       0.000E+00
#> 157 1.470981E-01   4       -1.490E-01          -2.820E-02      -3.547E+00
#> 158 1.470981E-01   1        0.000E+00           0.000E+00       0.000E+00
#> 159 1.566377E-01   4       -1.834E-01          -3.443E-02      -3.609E+00
#> 160 1.566377E-01   1        0.000E+00           0.000E+00       0.000E+00
#> 161 1.680853E-01   4       -2.246E-01          -4.122E-02      -3.600E+00
#> 162 1.680853E-01   1        0.000E+00           0.000E+00       0.000E+00
#> 163 1.818223E-01   4       -2.730E-01          -4.837E-02      -3.521E+00
#> 164 1.818223E-01   1        0.000E+00           0.000E+00       0.000E+00
#> 165 1.983068E-01   4       -3.285E-01          -5.557E-02      -3.371E+00
#> 166 1.983068E-01   1        0.000E+00           0.000E+00       0.000E+00
#> 167 2.180882E-01   4       -3.911E-01          -6.256E-02      -3.163E+00
#> 168 2.180882E-01   1        0.000E+00           0.000E+00       0.000E+00
#> 169 2.418258E-01   4       -4.601E-01          -6.901E-02      -2.907E+00
#> 170 2.418258E-01   1        0.000E+00           0.000E+00       0.000E+00
#> 171 2.703109E-01   4       -5.345E-01          -7.440E-02      -2.612E+00
#> 172 2.703109E-01   1        0.000E+00           0.000E+00       0.000E+00
#> 173 3.044931E-01   4       -6.128E-01          -7.824E-02      -2.289E+00
#> 174 3.044931E-01   1        0.000E+00           0.000E+00       0.000E+00
#> 175 3.455118E-01   4       -6.925E-01          -7.972E-02      -1.944E+00
#> 176 3.455118E-01   1        0.000E+00           0.000E+00       0.000E+00
#> 177 3.504340E-01   4       -7.019E-01          -9.382E-03      -1.906E+00
#> 178 3.504340E-01   1        0.000E+00           0.000E+00       0.000E+00
#> 179 3.563407E-01   4       -7.129E-01          -1.099E-02      -1.861E+00
#> 180 3.563407E-01   1        0.000E+00           0.000E+00       0.000E+00
#> 181 3.634287E-01   4       -7.257E-01          -1.280E-02      -1.806E+00
#> 182 3.634287E-01   1        0.000E+00           0.000E+00       0.000E+00
#> 183 3.719343E-01   4       -7.405E-01          -1.483E-02      -1.743E+00
#> 184 3.719343E-01   1        0.000E+00           0.000E+00       0.000E+00
#> 185 3.821411E-01   4       -7.575E-01          -1.704E-02      -1.669E+00
#> 186 3.821411E-01   1        0.000E+00           0.000E+00       0.000E+00
#> 187 3.943892E-01   4       -7.769E-01          -1.940E-02      -1.584E+00
#> 188 3.943892E-01   1        0.000E+00           0.000E+00       0.000E+00
#> 189 4.090869E-01   4       -8.025E-01          -2.556E-02      -1.739E+00
#> 190 4.090869E-01   1        0.000E+00           0.000E+00       0.000E+00
#> 191 4.267241E-01   4       -8.302E-01          -2.777E-02      -1.575E+00
#> 192 4.267241E-01   1        0.000E+00           0.000E+00       0.000E+00
#> 193 4.478888E-01   4       -8.603E-01          -3.007E-02      -1.421E+00
#> 194 4.478888E-01   1        0.000E+00           0.000E+00       0.000E+00
#> 195 4.732865E-01   4       -8.925E-01          -3.221E-02      -1.268E+00
#> 196 4.732865E-01   1        0.000E+00           0.000E+00       0.000E+00
#> 197 5.000000E-01   4       -9.228E-01          -3.024E-02      -1.132E+00
#> 198 5.000000E-01   1        0.000E+00           0.000E+00       0.000E+00
#> 199 5.320562E-01   4       -9.546E-01          -3.187E-02      -9.941E-01
#> 200 5.320562E-01   1        0.000E+00           0.000E+00       0.000E+00
#> 201 5.705236E-01   4       -9.876E-01          -3.295E-02      -8.565E-01
#> 202 5.705236E-01   1        0.000E+00           0.000E+00       0.000E+00
#> 203 6.166846E-01   4       -1.021E+00          -3.338E-02      -7.231E-01
#> 204 6.166846E-01   1        0.000E+00           0.000E+00       0.000E+00
#> 205 6.720777E-01   4       -1.054E+00          -3.310E-02      -5.976E-01
#> 206 6.720777E-01   1        0.000E+00           0.000E+00       0.000E+00
#> 207 7.385494E-01   4       -1.086E+00          -3.210E-02      -4.829E-01
#> 208 7.385494E-01   1        0.000E+00           0.000E+00       0.000E+00
#> 209 8.183155E-01   4       -1.117E+00          -3.043E-02      -3.815E-01
#> 210 8.183155E-01   1        0.000E+00           0.000E+00       0.000E+00
#> 211 9.140348E-01   4       -1.145E+00          -2.822E-02      -2.949E-01
#> 212 9.140348E-01   1        0.000E+00           0.000E+00       0.000E+00
#> 213 1.000000E+00   4       -1.165E+00          -2.039E-02      -2.371E-01
#> 214 1.000000E+00   1        0.000E+00           0.000E+00       0.000E+00
#> 215 1.103158E+00   4       -1.184E+00          -1.929E-02      -1.870E-01
#> 216 1.103158E+00   1        0.000E+00           0.000E+00       0.000E+00
#> 217 1.226948E+00   4       -1.202E+00          -1.794E-02      -1.449E-01
#> 218 1.226948E+00   1        0.000E+00           0.000E+00       0.000E+00
#> 219 1.375496E+00   4       -1.219E+00          -1.644E-02      -1.107E-01
#> 220 1.375496E+00   1        0.000E+00           0.000E+00       0.000E+00
#> 221 1.500000E+00   4       -1.230E+00          -1.118E-02      -8.982E-02
#> 222 1.500000E+00   1        0.000E+00           0.000E+00       0.000E+00
#> 223 1.649405E+00   4       -1.241E+00          -1.077E-02      -7.209E-02
#> 224 1.649405E+00   1        0.000E+00           0.000E+00       0.000E+00
#> 225 1.828690E+00   4       -1.251E+00          -1.029E-02      -5.741E-02
#> 226 1.828690E+00   1        0.000E+00           0.000E+00       0.000E+00
#> 227 2.000000E+00   4       -1.259E+00          -8.126E-03      -4.744E-02
#> 228 2.000000E+00   1        0.000E+00           0.000E+00       0.000E+00
#> 229 2.205572E+00   4       -1.267E+00          -8.009E-03      -3.896E-02
#> 230 2.205572E+00   1        0.000E+00           0.000E+00       0.000E+00
#> 231 2.452258E+00   4       -1.275E+00          -7.893E-03      -3.200E-02
#> 232 2.452258E+00   1        0.000E+00           0.000E+00       0.000E+00
#> 233 2.500000E+00   4       -1.277E+00          -1.472E-03      -3.084E-02
#> 234 2.500000E+00   1        0.000E+00           0.000E+00       0.000E+00
#> 235 2.557291E+00   4       -1.278E+00          -1.690E-03      -2.951E-02
#> 236 2.557291E+00   1        0.000E+00           0.000E+00       0.000E+00
#> 237 2.626039E+00   4       -1.280E+00          -1.927E-03      -2.803E-02
#> 238 2.626039E+00   1        0.000E+00           0.000E+00       0.000E+00
#> 239 2.708538E+00   4       -1.282E+00          -2.185E-03      -2.649E-02
#> 240 2.708538E+00   1        0.000E+00           0.000E+00       0.000E+00
#> 241 2.807536E+00   4       -1.285E+00          -2.462E-03      -2.487E-02
#> 242 2.807536E+00   1        0.000E+00           0.000E+00       0.000E+00
#> 243 2.926334E+00   4       -1.288E+00          -2.756E-03      -2.320E-02
#> 244 2.926334E+00   1        0.000E+00           0.000E+00       0.000E+00
#> 245 3.000000E+00   4       -1.289E+00          -1.640E-03      -2.226E-02
#> 246 3.000000E+00   1        0.000E+00           0.000E+00       0.000E+00
#> 247 3.088399E+00   4       -1.291E+00          -1.879E-03      -2.125E-02
#> 248 3.088399E+00   1        0.000E+00           0.000E+00       0.000E+00
#> 249 3.194478E+00   4       -1.293E+00          -2.141E-03      -2.018E-02
#> 250 3.194478E+00   1        0.000E+00           0.000E+00       0.000E+00
#> 251 3.321773E+00   4       -1.296E+00          -2.428E-03      -1.907E-02
#> 252 3.321773E+00   1        0.000E+00           0.000E+00       0.000E+00
#> 253 3.474527E+00   4       -1.298E+00          -2.740E-03      -1.794E-02
#> 254 3.474527E+00   1        0.000E+00           0.000E+00       0.000E+00
#> 255 3.500000E+00   4       -1.299E+00          -4.521E-04      -1.775E-02
#> 256 3.500000E+00   1        0.000E+00           0.000E+00       0.000E+00
#> 257 3.530567E+00   4       -1.299E+00          -5.361E-04      -1.754E-02
#> 258 3.530567E+00   1        0.000E+00           0.000E+00       0.000E+00
#> 259 3.567247E+00   4       -1.300E+00          -6.345E-04      -1.730E-02
#> 260 3.567247E+00   1        0.000E+00           0.000E+00       0.000E+00
#> 261 3.611264E+00   4       -1.301E+00          -7.494E-04      -1.703E-02
#> 262 3.611264E+00   1        0.000E+00           0.000E+00       0.000E+00
#> 263 3.664084E+00   4       -1.302E+00          -8.830E-04      -1.672E-02
#> 264 3.664084E+00   1        0.000E+00           0.000E+00       0.000E+00
#> 265 3.727468E+00   4       -1.303E+00          -1.038E-03      -1.637E-02
#> 266 3.727468E+00   1        0.000E+00           0.000E+00       0.000E+00
#> 267 3.803528E+00   4       -1.304E+00          -1.216E-03      -1.599E-02
#> 268 3.803528E+00   1        0.000E+00           0.000E+00       0.000E+00
#> 269 3.894801E+00   4       -1.305E+00          -1.421E-03      -1.557E-02
#> 270 3.894801E+00   1        0.000E+00           0.000E+00       0.000E+00
#> 271 4.000000E+00   4       -1.307E+00          -1.593E-03      -1.514E-02
#> 272 4.000000E+00   1        0.000E+00           0.000E+00       0.000E+00
#> 273 4.126239E+00   4       -1.309E+00          -1.855E-03      -1.469E-02
#> 274 4.126239E+00   1        0.000E+00           0.000E+00       0.000E+00
#> 275 4.277726E+00   4       -1.311E+00          -2.159E-03      -1.425E-02
#> 276 4.277726E+00   1        0.000E+00           0.000E+00       0.000E+00
#> 277 4.459510E+00   4       -1.314E+00          -2.514E-03      -1.383E-02
#> 278 4.459510E+00   1        0.000E+00           0.000E+00       0.000E+00
#> 279 4.500000E+00   4       -1.314E+00          -5.564E-04      -1.374E-02
#> 280 4.500000E+00   1        0.000E+00           0.000E+00       0.000E+00
#> 281 4.548588E+00   4       -1.315E+00          -6.630E-04      -1.365E-02
#> 282 4.548588E+00   1        0.000E+00           0.000E+00       0.000E+00
#> 283 4.606894E+00   4       -1.316E+00          -7.895E-04      -1.354E-02
#> 284 4.606894E+00   1        0.000E+00           0.000E+00       0.000E+00
#> 285 4.676861E+00   4       -1.316E+00          -9.397E-04      -1.343E-02
#> 286 4.676861E+00   1        0.000E+00           0.000E+00       0.000E+00
#> 287 4.760821E+00   4       -1.318E+00          -1.118E-03      -1.332E-02
#> 288 4.760821E+00   1        0.000E+00           0.000E+00       0.000E+00
#> 289 4.861573E+00   4       -1.319E+00          -1.330E-03      -1.320E-02
#> 290 4.861573E+00   1        0.000E+00           0.000E+00       0.000E+00
#> 291 4.982476E+00   4       -1.320E+00          -1.585E-03      -1.311E-02
#> 292 4.982476E+00   1        0.000E+00           0.000E+00       0.000E+00
#> 293 5.000000E+00   4       -1.321E+00          -2.294E-04      -1.309E-02
#> 294 5.000000E+00   1        0.000E+00           0.000E+00       0.000E+00
#> 295 5.021029E+00   4       -1.321E+00          -2.750E-04      -1.308E-02
#> 296 5.021029E+00   1        0.000E+00           0.000E+00       0.000E+00
#> 297 5.046264E+00   4       -1.321E+00          -3.295E-04      -1.306E-02
#> 298 5.046264E+00   1        0.000E+00           0.000E+00       0.000E+00
#> 299 5.076545E+00   4       -1.322E+00          -3.949E-04      -1.304E-02
#> 300 5.076545E+00   1        0.000E+00           0.000E+00       0.000E+00
#> 301 5.112883E+00   4       -1.322E+00          -4.732E-04      -1.302E-02
#> 302 5.112883E+00   1        0.000E+00           0.000E+00       0.000E+00
#> 303 5.156489E+00   4       -1.323E+00          -5.670E-04      -1.300E-02
#> 304 5.156489E+00   1        0.000E+00           0.000E+00       0.000E+00
#> 305 5.208815E+00   4       -1.323E+00          -6.796E-04      -1.299E-02
#> 306 5.208815E+00   1        0.000E+00           0.000E+00       0.000E+00
#> 307 5.271607E+00   4       -1.324E+00          -8.148E-04      -1.298E-02
#> 308 5.271607E+00   1        0.000E+00           0.000E+00       0.000E+00
#> 309 5.346957E+00   4       -1.325E+00          -9.778E-04      -1.298E-02
#> 310 5.346957E+00   1        0.000E+00           0.000E+00       0.000E+00
#> 311 5.437378E+00   4       -1.326E+00          -1.175E-03      -1.299E-02
#> 312 5.437378E+00   1        0.000E+00           0.000E+00       0.000E+00
#> 313 5.500000E+00   4       -1.327E+00          -8.146E-04      -1.301E-02
#> 314 5.500000E+00   1        0.000E+00           0.000E+00       0.000E+00
#> 315 5.575147E+00   4       -1.328E+00          -9.799E-04      -1.304E-02
#> 316 5.575147E+00   1        0.000E+00           0.000E+00       0.000E+00
#> 317 5.665322E+00   4       -1.329E+00          -1.181E-03      -1.309E-02
#> 318 5.665322E+00   1        0.000E+00           0.000E+00       0.000E+00
#> 319 5.773533E+00   4       -1.331E+00          -1.426E-03      -1.317E-02
#> 320 5.773533E+00   1        0.000E+00           0.000E+00       0.000E+00
#> 321 5.903387E+00   4       -1.333E+00          -1.727E-03      -1.330E-02
#> 322 5.903387E+00   1        0.000E+00           0.000E+00       0.000E+00
#> 323 6.000000E+00   4       -1.334E+00          -1.294E-03      -1.340E-02
#> 324 6.000000E+00   1        0.000E+00           0.000E+00       0.000E+00
#> 325 6.115936E+00   4       -1.335E+00          -1.570E-03      -1.354E-02
#> 326 6.115936E+00   1        0.000E+00           0.000E+00       0.000E+00
#> 327 6.255059E+00   4       -1.337E+00          -1.911E-03      -1.373E-02
#> 328 6.255059E+00   1        0.000E+00           0.000E+00       0.000E+00
#> 329 6.422007E+00   4       -1.340E+00          -2.336E-03      -1.400E-02
#> 330 6.422007E+00   1        0.000E+00           0.000E+00       0.000E+00
#> 331 6.500000E+00   4       -1.341E+00          -1.101E-03      -1.412E-02
#> 332 6.500000E+00   1        0.000E+00           0.000E+00       0.000E+00
#> 333 6.593591E+00   4       -1.342E+00          -1.337E-03      -1.428E-02
#> 334 6.593591E+00   1        0.000E+00           0.000E+00       0.000E+00
#> 335 6.705900E+00   4       -1.344E+00          -1.627E-03      -1.449E-02
#> 336 6.705900E+00   1        0.000E+00           0.000E+00       0.000E+00
#> 337 6.840672E+00   4       -1.346E+00          -1.988E-03      -1.475E-02
#> 338 6.840672E+00   1        0.000E+00           0.000E+00       0.000E+00
#> 339 7.000000E+00   4       -1.348E+00          -2.402E-03      -1.508E-02
#> 340 7.000000E+00   1        0.000E+00           0.000E+00       0.000E+00
#> 341 7.191194E+00   4       -1.351E+00          -2.963E-03      -1.550E-02
#> 342 7.191194E+00   1        0.000E+00           0.000E+00       0.000E+00
#> 343 7.420627E+00   4       -1.355E+00          -3.679E-03      -1.603E-02
#> 344 7.420627E+00   1        0.000E+00           0.000E+00       0.000E+00
#> 345 7.500000E+00   4       -1.356E+00          -1.288E-03      -1.623E-02
#> 346 7.500000E+00   1        0.000E+00           0.000E+00       0.000E+00
#> 347 7.595248E+00   4       -1.358E+00          -1.569E-03      -1.647E-02
#> 348 7.595248E+00   1        0.000E+00           0.000E+00       0.000E+00
#> 349 7.709545E+00   4       -1.360E+00          -1.918E-03      -1.678E-02
#> 350 7.709545E+00   1        0.000E+00           0.000E+00       0.000E+00
#> 351 7.846702E+00   4       -1.362E+00          -2.355E-03      -1.717E-02
#> 352 7.846702E+00   1        0.000E+00           0.000E+00       0.000E+00
#> 353 8.000000E+00   4       -1.365E+00          -2.704E-03      -1.764E-02
#> 354 8.000000E+00   1        0.000E+00           0.000E+00       0.000E+00
#> 355 8.183957E+00   4       -1.368E+00          -3.359E-03      -1.826E-02
#> 356 8.183957E+00   1        0.000E+00           0.000E+00       0.000E+00
#> 357 8.404706E+00   4       -1.372E+00          -4.204E-03      -1.904E-02
#> 358 8.404706E+00   1        0.000E+00           0.000E+00       0.000E+00
#> 359 8.500000E+00   4       -1.374E+00          -1.850E-03      -1.942E-02
#> 360 8.500000E+00   1        0.000E+00           0.000E+00       0.000E+00
#> 361 8.614353E+00   4       -1.376E+00          -2.275E-03      -1.989E-02
#> 362 8.614353E+00   1        0.000E+00           0.000E+00       0.000E+00
#> 363 8.751577E+00   4       -1.379E+00          -2.815E-03      -2.052E-02
#> 364 8.751577E+00   1        0.000E+00           0.000E+00       0.000E+00
#> 365 8.916245E+00   4       -1.383E+00          -3.501E-03      -2.126E-02
#> 366 8.916245E+00   1        0.000E+00           0.000E+00       0.000E+00
#> 367 9.000000E+00   4       -1.384E+00          -1.809E-03      -2.160E-02
#> 368 9.000000E+00   1        0.000E+00           0.000E+00       0.000E+00
#> 369 9.100506E+00   4       -1.387E+00          -2.209E-03      -2.198E-02
#> 370 9.100506E+00   1        0.000E+00           0.000E+00       0.000E+00
#> 371 9.221113E+00   4       -1.389E+00          -2.698E-03      -2.237E-02
#> 372 9.221113E+00   1        0.000E+00           0.000E+00       0.000E+00
#> 373 9.365842E+00   4       -1.393E+00          -3.288E-03      -2.272E-02
#> 374 9.365842E+00   1        0.000E+00           0.000E+00       0.000E+00
#> 375 9.500000E+00   4       -1.396E+00          -3.080E-03      -2.296E-02
#> 376 9.500000E+00   1        0.000E+00           0.000E+00       0.000E+00
#> 377 9.660990E+00   4       -1.399E+00          -3.726E-03      -2.315E-02
#> 378 9.660990E+00   1        0.000E+00           0.000E+00       0.000E+00
#> 379 9.854178E+00   4       -1.404E+00          -4.499E-03      -2.329E-02
#> 380 9.854178E+00   1        0.000E+00           0.000E+00       0.000E+00
#> 381 1.000000E+01   4       -1.407E+00          -3.407E-03      -2.336E-02
#> 382 1.000000E+01   1        0.000E+00           0.000E+00       0.000E+00
#> 383 1.010000E+01   4       -1.410E+00          -2.339E-03      -2.339E-02
#> 384 1.010000E+01   1        0.000E+00           0.000E+00       0.000E+00
#>     HEAT_FLUX_TOTAL HEAT_FLUX_TIMESTEP HEAT_FLUX_RATE
#> 1        -1.637E+01         -1.637E+01     -1.364E+05
#> 2        -1.637E+01         -1.637E+01     -1.364E+05
#> 3        -3.600E+01         -1.964E+01     -1.364E+05
#> 4        -3.600E+01         -1.964E+01     -1.364E+05
#> 5        -5.956E+01         -2.356E+01     -1.363E+05
#> 6        -5.956E+01         -2.356E+01     -1.363E+05
#> 7        -8.783E+01         -2.826E+01     -1.363E+05
#> 8        -8.783E+01         -2.826E+01     -1.363E+05
#> 9        -1.217E+02         -3.391E+01     -1.363E+05
#> 10       -1.217E+02         -3.391E+01     -1.363E+05
#> 11       -1.624E+02         -4.067E+01     -1.362E+05
#> 12       -1.624E+02         -4.067E+01     -1.362E+05
#> 13       -2.112E+02         -4.878E+01     -1.361E+05
#> 14       -2.112E+02         -4.878E+01     -1.361E+05
#> 15       -2.697E+02         -5.850E+01     -1.361E+05
#> 16       -2.697E+02         -5.850E+01     -1.361E+05
#> 17       -3.398E+02         -7.016E+01     -1.360E+05
#> 18       -3.398E+02         -7.016E+01     -1.360E+05
#> 19       -4.240E+02         -8.412E+01     -1.359E+05
#> 20       -4.240E+02         -8.412E+01     -1.359E+05
#> 21       -5.248E+02         -1.008E+02     -1.357E+05
#> 22       -5.248E+02         -1.008E+02     -1.357E+05
#> 23       -6.457E+02         -1.209E+02     -1.356E+05
#> 24       -6.457E+02         -1.209E+02     -1.356E+05
#> 25       -7.905E+02         -1.448E+02     -1.354E+05
#> 26       -7.905E+02         -1.448E+02     -1.354E+05
#> 27       -9.640E+02         -1.735E+02     -1.351E+05
#> 28       -9.640E+02         -1.735E+02     -1.351E+05
#> 29       -9.775E+02         -1.350E+01     -1.350E+05
#> 30       -9.775E+02         -1.350E+01     -1.350E+05
#> 31       -9.937E+02         -1.620E+01     -1.350E+05
#> 32       -9.937E+02         -1.620E+01     -1.350E+05
#> 33       -1.013E+03         -1.943E+01     -1.350E+05
#> 34       -1.013E+03         -1.943E+01     -1.350E+05
#> 35       -1.036E+03         -2.332E+01     -1.349E+05
#> 36       -1.036E+03         -2.332E+01     -1.349E+05
#> 37       -1.064E+03         -2.797E+01     -1.349E+05
#> 38       -1.064E+03         -2.797E+01     -1.349E+05
#> 39       -1.098E+03         -3.355E+01     -1.348E+05
#> 40       -1.098E+03         -3.355E+01     -1.348E+05
#> 41       -1.138E+03         -4.025E+01     -1.348E+05
#> 42       -1.138E+03         -4.025E+01     -1.348E+05
#> 43       -1.187E+03         -4.828E+01     -1.347E+05
#> 44       -1.187E+03         -4.828E+01     -1.347E+05
#> 45       -1.244E+03         -5.790E+01     -1.347E+05
#> 46       -1.244E+03         -5.790E+01     -1.347E+05
#> 47       -1.314E+03         -6.943E+01     -1.346E+05
#> 48       -1.314E+03         -6.943E+01     -1.346E+05
#> 49       -1.397E+03         -8.325E+01     -1.345E+05
#> 50       -1.397E+03         -8.325E+01     -1.345E+05
#> 51       -1.497E+03         -9.981E+01     -1.343E+05
#> 52       -1.497E+03         -9.981E+01     -1.343E+05
#> 53       -1.510E+03         -1.342E+01     -1.342E+05
#> 54       -1.510E+03         -1.342E+01     -1.342E+05
#> 55       -1.526E+03         -1.611E+01     -1.342E+05
#> 56       -1.526E+03         -1.611E+01     -1.342E+05
#> 57       -1.546E+03         -1.933E+01     -1.342E+05
#> 58       -1.546E+03         -1.933E+01     -1.342E+05
#> 59       -1.569E+03         -2.319E+01     -1.342E+05
#> 60       -1.569E+03         -2.319E+01     -1.342E+05
#> 61       -1.597E+03         -2.782E+01     -1.341E+05
#> 62       -1.597E+03         -2.782E+01     -1.341E+05
#> 63       -1.630E+03         -3.337E+01     -1.341E+05
#> 64       -1.630E+03         -3.337E+01     -1.341E+05
#> 65       -1.670E+03         -4.003E+01     -1.340E+05
#> 66       -1.670E+03         -4.003E+01     -1.340E+05
#> 67       -1.718E+03         -4.801E+01     -1.340E+05
#> 68       -1.718E+03         -4.801E+01     -1.340E+05
#> 69       -1.776E+03         -5.758E+01     -1.339E+05
#> 70       -1.776E+03         -5.758E+01     -1.339E+05
#> 71       -1.845E+03         -6.905E+01     -1.338E+05
#> 72       -1.845E+03         -6.905E+01     -1.338E+05
#> 73       -1.928E+03         -8.279E+01     -1.337E+05
#> 74       -1.928E+03         -8.279E+01     -1.337E+05
#> 75       -2.027E+03         -9.926E+01     -1.336E+05
#> 76       -2.027E+03         -9.926E+01     -1.336E+05
#> 77       -2.146E+03         -1.190E+02     -1.334E+05
#> 78       -2.146E+03         -1.190E+02     -1.334E+05
#> 79       -2.288E+03         -1.426E+02     -1.333E+05
#> 80       -2.288E+03         -1.426E+02     -1.333E+05
#> 81       -2.459E+03         -1.708E+02     -1.330E+05
#> 82       -2.459E+03         -1.708E+02     -1.330E+05
#> 83       -2.664E+03         -2.046E+02     -1.328E+05
#> 84       -2.664E+03         -2.046E+02     -1.328E+05
#> 85       -2.909E+03         -2.450E+02     -1.325E+05
#> 86       -2.909E+03         -2.450E+02     -1.325E+05
#> 87       -3.202E+03         -2.932E+02     -1.321E+05
#> 88       -3.202E+03         -2.932E+02     -1.321E+05
#> 89       -3.553E+03         -3.507E+02     -1.317E+05
#> 90       -3.553E+03         -3.506E+02     -1.317E+05
#> 91       -3.972E+03         -4.192E+02     -1.312E+05
#> 92       -3.972E+03         -4.192E+02     -1.312E+05
#> 93       -4.473E+03         -5.008E+02     -1.306E+05
#> 94       -4.472E+03         -5.007E+02     -1.306E+05
#> 95       -5.070E+03         -5.977E+02     -1.299E+05
#> 96       -5.070E+03         -5.977E+02     -1.299E+05
#> 97       -5.783E+03         -7.129E+02     -1.291E+05
#> 98       -5.783E+03         -7.128E+02     -1.291E+05
#> 99       -6.632E+03         -8.494E+02     -1.282E+05
#> 100      -6.632E+03         -8.491E+02     -1.282E+05
#> 101      -7.645E+03         -1.012E+03     -1.273E+05
#> 102      -7.642E+03         -1.010E+03     -1.271E+05
#> 103      -9.396E+03         -1.751E+03     -1.836E+05
#> 104      -8.852E+03         -1.209E+03     -1.268E+05
#> 105      -3.115E+04         -2.175E+04     -1.900E+06
#> 106      -1.059E+04         -1.740E+03     -1.520E+05
#> 107      -2.178E+05         -1.867E+05     -1.359E+07
#> 108      -1.496E+04         -4.363E+03     -3.176E+05
#> 109      -4.514E+05         -2.336E+05     -3.330E+07
#> 110      -1.904E+04         -4.081E+03     -5.816E+05
#> 111      -4.567E+05         -5.245E+03     -4.371E+07
#> 112      -1.912E+04         -8.587E+01     -7.156E+05
#> 113      -4.630E+05         -6.344E+03     -4.406E+07
#> 114      -1.923E+04         -1.037E+02     -7.203E+05
#> 115      -4.708E+05         -7.738E+03     -4.478E+07
#> 116      -1.935E+04         -1.261E+02     -7.299E+05
#> 117      -4.802E+05         -9.452E+03     -4.558E+07
#> 118      -1.950E+04         -1.535E+02     -7.401E+05
#> 119      -4.918E+05         -1.157E+04     -4.649E+07
#> 120      -1.969E+04         -1.870E+02     -7.514E+05
#> 121      -5.060E+05         -1.419E+04     -4.753E+07
#> 122      -1.992E+04         -2.281E+02     -7.640E+05
#> 123      -5.234E+05         -1.746E+04     -4.873E+07
#> 124      -2.020E+04         -2.789E+02     -7.783E+05
#> 125      -5.450E+05         -2.155E+04     -5.013E+07
#> 126      -2.054E+04         -3.416E+02     -7.944E+05
#> 127      -5.717E+05         -2.671E+04     -5.177E+07
#> 128      -2.096E+04         -4.195E+02     -8.131E+05
#> 129      -6.049E+05         -3.325E+04     -5.370E+07
#> 130      -2.148E+04         -5.167E+02     -8.345E+05
#> 131      -6.465E+05         -4.160E+04     -5.598E+07
#> 132      -2.212E+04         -6.385E+02     -8.593E+05
#> 133      -6.989E+05         -5.231E+04     -5.867E+07
#> 134      -2.291E+04         -7.915E+02     -8.877E+05
#> 135      -7.650E+05         -6.615E+04     -6.183E+07
#> 136      -2.389E+04         -9.844E+02     -9.201E+05
#> 137      -8.492E+05         -8.415E+04     -6.554E+07
#> 138      -2.512E+04         -1.228E+03     -9.568E+05
#> 139      -9.568E+05         -1.077E+05     -6.989E+07
#> 140      -2.666E+04         -1.538E+03     -9.986E+05
#> 141      -1.095E+06         -1.386E+05     -7.495E+07
#> 142      -2.859E+04         -1.932E+03     -1.045E+06
#> 143      -1.275E+06         -1.792E+05     -8.076E+07
#> 144      -3.102E+04         -2.433E+03     -1.097E+06
#> 145      -1.507E+06         -2.326E+05     -8.735E+07
#> 146      -3.409E+04         -3.068E+03     -1.152E+06
#> 147      -1.810E+06         -3.024E+05     -9.465E+07
#> 148      -3.797E+04         -3.873E+03     -1.212E+06
#> 149      -2.203E+06         -3.930E+05     -1.025E+08
#> 150      -4.285E+04         -4.889E+03     -1.275E+06
#> 151      -2.712E+06         -5.090E+05     -1.106E+08
#> 152      -4.902E+04         -6.164E+03     -1.340E+06
#> 153      -3.366E+06         -6.546E+05     -1.186E+08
#> 154      -5.677E+04         -7.754E+03     -1.404E+06
#> 155      -4.199E+06         -8.329E+05     -1.257E+08
#> 156      -6.648E+04         -9.713E+03     -1.466E+06
#> 157      -5.244E+06         -1.045E+06     -1.314E+08
#> 158      -7.858E+04         -1.210E+04     -1.522E+06
#> 159      -6.534E+06         -1.290E+06     -1.352E+08
#> 160      -9.355E+04         -1.496E+04     -1.569E+06
#> 161      -8.106E+06         -1.572E+06     -1.373E+08
#> 162      -1.119E+05         -1.834E+04     -1.603E+06
#> 163      -9.980E+06         -1.874E+06     -1.364E+08
#> 164      -1.341E+05         -2.223E+04     -1.619E+06
#> 165      -1.216E+07         -2.177E+06     -1.321E+08
#> 166      -1.609E+05         -2.676E+04     -1.623E+06
#> 167      -1.464E+07         -2.487E+06     -1.257E+08
#> 168      -1.933E+05         -3.239E+04     -1.637E+06
#> 169      -1.743E+07         -2.788E+06     -1.175E+08
#> 170      -2.330E+05         -3.968E+04     -1.672E+06
#> 171      -2.050E+07         -3.066E+06     -1.076E+08
#> 172      -2.822E+05         -4.920E+04     -1.727E+06
#> 173      -2.380E+07         -3.298E+06     -9.648E+07
#> 174      -3.437E+05         -6.156E+04     -1.801E+06
#> 175      -2.725E+07         -3.454E+06     -8.420E+07
#> 176      -4.208E+05         -7.711E+04     -1.880E+06
#> 177      -2.763E+07         -3.794E+05     -7.708E+07
#> 178      -4.303E+05         -9.460E+03     -1.922E+06
#> 179      -2.808E+07         -4.460E+05     -7.551E+07
#> 180      -4.417E+05         -1.140E+04     -1.930E+06
#> 181      -2.860E+07         -5.218E+05     -7.361E+07
#> 182      -4.554E+05         -1.373E+04     -1.938E+06
#> 183      -2.920E+07         -6.070E+05     -7.136E+07
#> 184      -4.720E+05         -1.655E+04     -1.946E+06
#> 185      -2.991E+07         -7.014E+05     -6.872E+07
#> 186      -4.919E+05         -1.994E+04     -1.953E+06
#> 187      -3.071E+07         -8.038E+05     -6.563E+07
#> 188      -5.159E+05         -2.400E+04     -1.959E+06
#> 189      -3.170E+07         -9.949E+05     -6.769E+07
#> 190      -5.430E+05         -2.709E+04     -1.843E+06
#> 191      -3.291E+07         -1.210E+06     -6.860E+07
#> 192      -5.724E+05         -2.945E+04     -1.670E+06
#> 193      -3.424E+07         -1.330E+06     -6.283E+07
#> 194      -6.057E+05         -3.323E+04     -1.570E+06
#> 195      -3.569E+07         -1.444E+06     -5.685E+07
#> 196      -6.433E+05         -3.766E+04     -1.483E+06
#> 197      -3.705E+07         -1.365E+06     -5.110E+07
#> 198      -6.806E+05         -3.727E+04     -1.395E+06
#> 199      -3.851E+07         -1.460E+06     -4.554E+07
#> 200      -7.224E+05         -4.176E+04     -1.303E+06
#> 201      -4.005E+07         -1.534E+06     -3.987E+07
#> 202      -7.685E+05         -4.614E+04     -1.200E+06
#> 203      -4.163E+07         -1.579E+06     -3.421E+07
#> 204      -8.187E+05         -5.021E+04     -1.088E+06
#> 205      -4.322E+07         -1.592E+06     -2.874E+07
#> 206      -8.725E+05         -5.378E+04     -9.708E+05
#> 207      -4.479E+07         -1.569E+06     -2.360E+07
#> 208      -9.292E+05         -5.673E+04     -8.534E+05
#> 209      -4.630E+07         -1.511E+06     -1.894E+07
#> 210      -9.883E+05         -5.905E+04     -7.403E+05
#> 211      -4.772E+07         -1.422E+06     -1.486E+07
#> 212      -1.049E+06         -6.085E+04     -6.357E+05
#> 213      -4.873E+07         -1.006E+06     -1.171E+07
#> 214      -1.097E+06         -4.740E+04     -5.514E+05
#> 215      -4.969E+07         -9.642E+05     -9.347E+06
#> 216      -1.147E+06         -5.010E+04     -4.857E+05
#> 217      -5.060E+07         -9.073E+05     -7.329E+06
#> 218      -1.199E+06         -5.286E+04     -4.270E+05
#> 219      -5.144E+07         -8.406E+05     -5.659E+06
#> 220      -1.255E+06         -5.585E+04     -3.760E+05
#> 221      -5.199E+07         -5.549E+05     -4.457E+06
#> 222      -1.297E+06         -4.202E+04     -3.375E+05
#> 223      -5.253E+07         -5.403E+05     -3.616E+06
#> 224      -1.344E+06         -4.616E+04     -3.089E+05
#> 225      -5.306E+07         -5.219E+05     -2.911E+06
#> 226      -1.394E+06         -5.076E+04     -2.831E+05
#> 227      -5.346E+07         -4.070E+05     -2.376E+06
#> 228      -1.439E+06         -4.487E+04     -2.619E+05
#> 229      -5.387E+07         -4.061E+05     -1.976E+06
#> 230      -1.489E+06         -5.028E+04     -2.446E+05
#> 231      -5.427E+07         -4.048E+05     -1.641E+06
#> 232      -1.546E+06         -5.637E+04     -2.285E+05
#> 233      -5.434E+07         -6.993E+04     -1.465E+06
#> 234      -1.556E+06         -1.048E+04     -2.195E+05
#> 235      -5.442E+07         -8.084E+04     -1.411E+06
#> 236      -1.569E+06         -1.242E+04     -2.167E+05
#> 237      -5.452E+07         -9.286E+04     -1.351E+06
#> 238      -1.583E+06         -1.468E+04     -2.136E+05
#> 239      -5.462E+07         -1.061E+05     -1.286E+06
#> 240      -1.601E+06         -1.733E+04     -2.100E+05
#> 241      -5.474E+07         -1.205E+05     -1.217E+06
#> 242      -1.621E+06         -2.040E+04     -2.061E+05
#> 243      -5.488E+07         -1.362E+05     -1.146E+06
#> 244      -1.645E+06         -2.398E+04     -2.019E+05
#> 245      -5.496E+07         -8.029E+04     -1.090E+06
#> 246      -1.660E+06         -1.461E+04     -1.984E+05
#> 247      -5.505E+07         -9.262E+04     -1.048E+06
#> 248      -1.677E+06         -1.730E+04     -1.957E+05
#> 249      -5.516E+07         -1.064E+05     -1.003E+06
#> 250      -1.697E+06         -2.044E+04     -1.927E+05
#> 251      -5.528E+07         -1.216E+05     -9.552E+05
#> 252      -1.722E+06         -2.411E+04     -1.894E+05
#> 253      -5.542E+07         -1.385E+05     -9.064E+05
#> 254      -1.750E+06         -2.840E+04     -1.859E+05
#> 255      -5.544E+07         -2.236E+04     -8.776E+05
#> 256      -1.755E+06         -4.681E+03     -1.838E+05
#> 257      -5.547E+07         -2.656E+04     -8.690E+05
#> 258      -1.760E+06         -5.598E+03     -1.831E+05
#> 259      -5.550E+07         -3.152E+04     -8.593E+05
#> 260      -1.767E+06         -6.691E+03     -1.824E+05
#> 261      -5.554E+07         -3.733E+04     -8.481E+05
#> 262      -1.775E+06         -7.992E+03     -1.816E+05
#> 263      -5.558E+07         -4.413E+04     -8.355E+05
#> 264      -1.784E+06         -9.538E+03     -1.806E+05
#> 265      -5.563E+07         -5.205E+04     -8.212E+05
#> 266      -1.796E+06         -1.137E+04     -1.794E+05
#> 267      -5.569E+07         -6.125E+04     -8.052E+05
#> 268      -1.809E+06         -1.355E+04     -1.781E+05
#> 269      -5.577E+07         -7.189E+04     -7.877E+05
#> 270      -1.825E+06         -1.612E+04     -1.766E+05
#> 271      -5.585E+07         -8.090E+04     -7.691E+05
#> 272      -1.844E+06         -1.841E+04     -1.750E+05
#> 273      -5.594E+07         -9.464E+04     -7.497E+05
#> 274      -1.866E+06         -2.187E+04     -1.732E+05
#> 275      -5.605E+07         -1.106E+05     -7.299E+05
#> 276      -1.892E+06         -2.594E+04     -1.713E+05
#> 277      -5.618E+07         -1.291E+05     -7.104E+05
#> 278      -1.922E+06         -3.075E+04     -1.692E+05
#> 279      -5.621E+07         -2.830E+04     -6.988E+05
#> 280      -1.929E+06         -6.796E+03     -1.678E+05
#> 281      -5.624E+07         -3.375E+04     -6.946E+05
#> 282      -1.937E+06         -8.132E+03     -1.674E+05
#> 283      -5.628E+07         -4.023E+04     -6.900E+05
#> 284      -1.947E+06         -9.725E+03     -1.668E+05
#> 285      -5.633E+07         -4.793E+04     -6.851E+05
#> 286      -1.959E+06         -1.162E+04     -1.661E+05
#> 287      -5.639E+07         -5.707E+04     -6.797E+05
#> 288      -1.973E+06         -1.389E+04     -1.654E+05
#> 289      -5.646E+07         -6.794E+04     -6.743E+05
#> 290      -1.989E+06         -1.658E+04     -1.645E+05
#> 291      -5.654E+07         -8.089E+04     -6.691E+05
#> 292      -2.009E+06         -1.978E+04     -1.636E+05
#> 293      -5.655E+07         -1.167E+04     -6.661E+05
#> 294      -2.012E+06         -2.857E+03     -1.630E+05
#> 295      -5.656E+07         -1.399E+04     -6.654E+05
#> 296      -2.015E+06         -3.425E+03     -1.629E+05
#> 297      -5.658E+07         -1.677E+04     -6.645E+05
#> 298      -2.019E+06         -4.105E+03     -1.627E+05
#> 299      -5.660E+07         -2.009E+04     -6.635E+05
#> 300      -2.024E+06         -4.920E+03     -1.625E+05
#> 301      -5.662E+07         -2.407E+04     -6.625E+05
#> 302      -2.030E+06         -5.895E+03     -1.622E+05
#> 303      -5.665E+07         -2.884E+04     -6.614E+05
#> 304      -2.037E+06         -7.061E+03     -1.619E+05
#> 305      -5.669E+07         -3.455E+04     -6.603E+05
#> 306      -2.046E+06         -8.456E+03     -1.616E+05
#> 307      -5.673E+07         -4.140E+04     -6.593E+05
#> 308      -2.056E+06         -1.012E+04     -1.612E+05
#> 309      -5.678E+07         -4.962E+04     -6.585E+05
#> 310      -2.068E+06         -1.211E+04     -1.607E+05
#> 311      -5.684E+07         -5.951E+04     -6.581E+05
#> 312      -2.082E+06         -1.449E+04     -1.602E+05
#> 313      -5.688E+07         -4.122E+04     -6.582E+05
#> 314      -2.092E+06         -1.001E+04     -1.598E+05
#> 315      -5.693E+07         -4.949E+04     -6.586E+05
#> 316      -2.104E+06         -1.198E+04     -1.594E+05
#> 317      -5.699E+07         -5.949E+04     -6.597E+05
#> 318      -2.119E+06         -1.433E+04     -1.590E+05
#> 319      -5.706E+07         -7.159E+04     -6.616E+05
#> 320      -2.136E+06         -1.715E+04     -1.585E+05
#> 321      -5.715E+07         -8.633E+04     -6.648E+05
#> 322      -2.156E+06         -2.051E+04     -1.579E+05
#> 323      -5.721E+07         -6.459E+04     -6.685E+05
#> 324      -2.172E+06         -1.521E+04     -1.574E+05
#> 325      -5.729E+07         -7.799E+04     -6.727E+05
#> 326      -2.190E+06         -1.820E+04     -1.570E+05
#> 327      -5.738E+07         -9.440E+04     -6.785E+05
#> 328      -2.212E+06         -2.178E+04     -1.565E+05
#> 329      -5.750E+07         -1.146E+05     -6.866E+05
#> 330      -2.238E+06         -2.605E+04     -1.560E+05
#> 331      -5.755E+07         -5.409E+04     -6.935E+05
#> 332      -2.250E+06         -1.214E+04     -1.557E+05
#> 333      -5.762E+07         -6.539E+04     -6.987E+05
#> 334      -2.264E+06         -1.455E+04     -1.554E+05
#> 335      -5.770E+07         -7.922E+04     -7.053E+05
#> 336      -2.282E+06         -1.743E+04     -1.552E+05
#> 337      -5.779E+07         -9.622E+04     -7.139E+05
#> 338      -2.303E+06         -2.087E+04     -1.549E+05
#> 339      -5.791E+07         -1.155E+05     -7.248E+05
#> 340      -2.327E+06         -2.463E+04     -1.546E+05
#> 341      -5.805E+07         -1.412E+05     -7.387E+05
#> 342      -2.357E+06         -2.950E+04     -1.543E+05
#> 343      -5.822E+07         -1.736E+05     -7.566E+05
#> 344      -2.392E+06         -3.533E+04     -1.540E+05
#> 345      -5.828E+07         -6.115E+04     -7.704E+05
#> 346      -2.404E+06         -1.221E+04     -1.538E+05
#> 347      -5.836E+07         -7.417E+04     -7.787E+05
#> 348      -2.419E+06         -1.464E+04     -1.537E+05
#> 349      -5.845E+07         -9.019E+04     -7.891E+05
#> 350      -2.436E+06         -1.756E+04     -1.536E+05
#> 351      -5.856E+07         -1.101E+05     -8.024E+05
#> 352      -2.458E+06         -2.106E+04     -1.536E+05
#> 353      -5.868E+07         -1.255E+05     -8.188E+05
#> 354      -2.481E+06         -2.353E+04     -1.535E+05
#> 355      -5.884E+07         -1.545E+05     -8.397E+05
#> 356      -2.509E+06         -2.825E+04     -1.535E+05
#> 357      -5.903E+07         -1.914E+05     -8.669E+05
#> 358      -2.543E+06         -3.392E+04     -1.537E+05
#> 359      -5.911E+07         -8.475E+04     -8.894E+05
#> 360      -2.558E+06         -1.466E+04     -1.538E+05
#> 361      -5.922E+07         -1.036E+05     -9.060E+05
#> 362      -2.576E+06         -1.760E+04     -1.539E+05
#> 363      -5.935E+07         -1.273E+05     -9.276E+05
#> 364      -2.597E+06         -2.116E+04     -1.542E+05
#> 365      -5.950E+07         -1.572E+05     -9.546E+05
#> 366      -2.622E+06         -2.545E+04     -1.546E+05
#> 367      -5.958E+07         -8.176E+04     -9.762E+05
#> 368      -2.635E+06         -1.297E+04     -1.549E+05
#> 369      -5.968E+07         -9.956E+04     -9.906E+05
#> 370      -2.651E+06         -1.559E+04     -1.551E+05
#> 371      -5.981E+07         -1.213E+05     -1.006E+06
#> 372      -2.669E+06         -1.875E+04     -1.555E+05
#> 373      -5.995E+07         -1.478E+05     -1.021E+06
#> 374      -2.692E+06         -2.256E+04     -1.559E+05
#> 375      -6.009E+07         -1.386E+05     -1.033E+06
#> 376      -2.713E+06         -2.096E+04     -1.563E+05
#> 377      -6.026E+07         -1.678E+05     -1.042E+06
#> 378      -2.738E+06         -2.522E+04     -1.566E+05
#> 379      -6.046E+07         -2.028E+05     -1.050E+06
#> 380      -2.769E+06         -3.034E+04     -1.571E+05
#> 381      -6.062E+07         -1.538E+05     -1.055E+06
#> 382      -2.791E+06         -2.296E+04     -1.574E+05
#> 383      -6.072E+07         -1.057E+05     -1.057E+06
#> 384      -2.807E+06         -1.577E+04     -1.577E+05