Run VS2dh model

Run VS2dh model

vs2di.run(engine = "vs2dh",
  engineDirectoryWin = this_extdata_file("engine/win"),
  engineDirectoryLinux = this_extdata_file("engine/linux"),
  model.path = this_extdata_file("vs2dh_example/tutorial2"),
  returnOutput = TRUE, showWarnings = TRUE, openTargetDir = FALSE,
  dbg = TRUE)

Arguments

engine
model engine 'vs2dh' (for flow & heat modelling) or 'vs2dt' (for flow & solute transport) (Default: "vs2dh")

engineDirectoryWin
default directory on Windows OS containing vs2dh3_3.exe and vs2dt3_3.exe (Default: system.file("extdata/engine/win", package = "kwb.vs2dh"))

engineDirectoryLinux
default directory on Linux OS containing vs2dh3_3.exe and vs2dt3_3.exe (Default: system.file("extdata/engine/linux", package = "kwb.vs2dh"))

model.path
full path to folder containing vs2dh.dat (default: system.file("extdata", "vs2dh_example/tutorial2", package = "kwb.vs2dh"))

returnOutput
if TRUE model output files variables.out, balance.out and obsPoints.out are imported into R completely; vs2dh.out (only warnings imported)

showWarnings
if TRUE print warning messages during simulation on screen (default: TRUE)

openTargetDir
If TRUE path containing model files will be opened in explorer (Default: FALSE)

dbg
if true text output on screen on model run progress

Value

Run VS2dh model and saves model output files in folder model.path. If returnOutput is TRUE the output is imported in R object

Examples

### Running model with default model.path and engine.path: 
model.path <- system.file("extdata", "vs2dh_example/tutorial2", package = "kwb.vs2dh")
res <- vs2di.run(model.path = model.path)

#> 1. Step: Create vs2dh.fil...Done!
#> 
#> 
#> 2. Step: Run model...finished after  0.0 seconds
#> 
#> 4. Step: Import model results...Done!
#> 
#> Print warnings during simulation (vs2dh.out):
#> Line:     28 :::      STOP SOLUTION IF MAXIMUM NO. OF ITERATIONS EXCEEDED IN ANY TIME STEP? T
#>  Line:    212 :::       DATA FOR RECHARGE PERIOD         1
#>  Line:    228 :::  ******** WARNING --- NODE TYPE OF 0 ASSIGNED TO BOUNDARY NODE    33    2
#>  Line:    229 :::  ******** WARNING --- NODE TYPE OF 0 ASSIGNED TO BOUNDARY NODE    33    3
#>  Line:    230 :::  ******** WARNING --- NODE TYPE OF 0 ASSIGNED TO BOUNDARY NODE    33    4
#>  Line:    231 :::  ******** WARNING --- NODE TYPE OF 0 ASSIGNED TO BOUNDARY NODE    33    5
#>  Line:    232 :::  ******** WARNING --- NODE TYPE OF 0 ASSIGNED TO BOUNDARY NODE    33    6
#>  Line:    233 :::  ******** WARNING --- NODE TYPE OF 0 ASSIGNED TO BOUNDARY NODE    33    7
#>  Line:    234 :::  ******** WARNING --- NODE TYPE OF 0 ASSIGNED TO BOUNDARY NODE    33    8
#>  Line:    235 :::  ******** WARNING --- NODE TYPE OF 0 ASSIGNED TO BOUNDARY NODE    33    9
#>  Line:    236 :::  ******** WARNING --- NODE TYPE OF 0 ASSIGNED TO BOUNDARY NODE    33   10
#>  Line:    237 :::  ******** WARNING --- NODE TYPE OF 0 ASSIGNED TO BOUNDARY NODE    33   11
#>  Line:    238 :::  ******** WARNING --- NODE TYPE OF 0 ASSIGNED TO BOUNDARY NODE    33   12
#>  Line:    239 :::  ******** WARNING --- NODE TYPE OF 0 ASSIGNED TO BOUNDARY NODE    33   13
#>  Line:    240 :::  ******** WARNING --- NODE TYPE OF 0 ASSIGNED TO BOUNDARY NODE    33   14
#>  Line:    241 :::  ******** WARNING --- NODE TYPE OF 0 ASSIGNED TO BOUNDARY NODE    33   15
#>  Line:    242 :::  ******** WARNING --- NODE TYPE OF 0 ASSIGNED TO BOUNDARY NODE    33   16
#>  Line:    243 :::  ******** WARNING --- NODE TYPE OF 0 ASSIGNED TO BOUNDARY NODE    33   17
#>  Line:    244 :::  ******** WARNING --- NODE TYPE OF 0 ASSIGNED TO BOUNDARY NODE    33   18
#>  Line:    245 :::  ******** WARNING --- NODE TYPE OF 0 ASSIGNED TO BOUNDARY NODE    33   19
#>  Line:    246 :::  ******** WARNING --- NODE TYPE OF 0 ASSIGNED TO BOUNDARY NODE    33   20
#>  Line:    247 :::  ******** WARNING --- NODE TYPE OF 0 ASSIGNED TO BOUNDARY NODE    33   21
#>  Line:    248 :::  ******** WARNING --- NODE TYPE OF 0 ASSIGNED TO BOUNDARY NODE    33   22
#>  Line:    249 :::  ******** WARNING --- NODE TYPE OF 0 ASSIGNED TO BOUNDARY NODE    33   23
#>  Line:    250 :::  ******** WARNING --- NODE TYPE OF 0 ASSIGNED TO BOUNDARY NODE    33   24
#>  Line:    251 :::  ******** WARNING --- NODE TYPE OF 0 ASSIGNED TO BOUNDARY NODE    33   25
#>  Line:    252 :::  ******** WARNING --- NODE TYPE OF 0 ASSIGNED TO BOUNDARY NODE    33   26
#>  Line:    253 :::  ******** WARNING --- NODE TYPE OF 0 ASSIGNED TO BOUNDARY NODE    33   27
#>  Line:    254 :::  ******** WARNING --- NODE TYPE OF 0 ASSIGNED TO BOUNDARY NODE    33   28
#>  Line:    255 :::  ******** WARNING --- NODE TYPE OF 0 ASSIGNED TO BOUNDARY NODE    33   29
#>  Line:    256 :::  ******** WARNING --- NODE TYPE OF 0 ASSIGNED TO BOUNDARY NODE    33   30
#>  Line:    257 :::  ******** WARNING --- NODE TYPE OF 0 ASSIGNED TO BOUNDARY NODE    33   31
#>  Line:    258 :::  ******** WARNING --- NODE TYPE OF 0 ASSIGNED TO BOUNDARY NODE    33   32
#>  Line:    259 :::  ******** WARNING --- NODE TYPE OF 0 ASSIGNED TO BOUNDARY NODE    33   33
#>  Line:    260 :::  ******** WARNING --- NODE TYPE OF 0 ASSIGNED TO BOUNDARY NODE    33   34
#>  Line:    261 :::  ******** WARNING --- NODE TYPE OF 0 ASSIGNED TO BOUNDARY NODE    33   35
#>  Line:    262 :::  ******** WARNING --- NODE TYPE OF 0 ASSIGNED TO BOUNDARY NODE    33   36
#>  Line:    263 :::  ******** WARNING --- NODE TYPE OF 0 ASSIGNED TO BOUNDARY NODE    33   37
#>  Line:    264 :::  ******** WARNING --- NODE TYPE OF 0 ASSIGNED TO BOUNDARY NODE    33   38
#>  Line:    265 :::  ******** WARNING --- NODE TYPE OF 0 ASSIGNED TO BOUNDARY NODE    33   39
#>  Line:    266 :::  ******** WARNING --- NODE TYPE OF 0 ASSIGNED TO BOUNDARY NODE    33   40
#>  Line:    267 :::  ******** WARNING --- NODE TYPE OF 0 ASSIGNED TO BOUNDARY NODE    33   41
#>  Line:    268 :::  ******** WARNING --- NODE TYPE OF 0 ASSIGNED TO BOUNDARY NODE    33   42
#>  Line:    269 :::  ******** WARNING --- NODE TYPE OF 0 ASSIGNED TO BOUNDARY NODE    33   43
#>  Line:    270 :::  ******** WARNING --- NODE TYPE OF 0 ASSIGNED TO BOUNDARY NODE    33   44
#>  Line:    271 :::  ******** WARNING --- NODE TYPE OF 0 ASSIGNED TO BOUNDARY NODE    33   45
#>  Line:    330 :::      EXCEEDED PERMITTED NUMBER OF ITERATIONS  ( =      201)
#>  Line:    336 :::      EXCEEDED PERMITTED NUMBER OF ITERATIONS  ( =      201)
#>  Line:    354 :::      EXCEEDED PERMITTED NUMBER OF ITERATIONS  ( =      201)
#>  Line:    506 :::       DATA FOR RECHARGE PERIOD         2
#>  Line:    599 :::      EXCEEDED PERMITTED NUMBER OF ITERATIONS  ( =      201)